+++ src/components/ConfirmModal.vue
@@ -29,5 +29,6 @@ defineProps({
     required: false
   }
 });
+
 const confirmModal = false;
 </script>
