+++ src/views/Active.vue
@@ -22,7 +22,11 @@
             {{ $t("deleteAll") }}
           </v-btn>
           <v-col cols="12">
-            <task-showing :tasks="tasksStore.activeTasks" />
+            <task-card
+              v-for="task in tasksStore.activeTasks"
+              :key="task.id"
+              :task="task"
+            />
           </v-col>
         </v-row>
       </v-container>
@@ -86,7 +90,7 @@
 
 <script lang="ts" setup>
 import confirmModal from "../components/ConfirmModal.vue";
-import TaskShowing from "../components/TaskList.vue";
+import TaskCard from "../components/TaskCard.vue";
 import { useTasksStore } from "../stores/tasks";
 
 const tasksStore = useTasksStore();
